vector-memory/requirements.txt
qdrant-client>=1.7.0
openai>=2.0.0
python-dotenv>=1.0.0
